doc: docstrings
