+++ SDeMo/src/pipelines/univariatetransforms.jl
@@ -1,13 +1,24 @@
-# No transformation
+"""
+    RawData
+
+A transformer that does *nothing* to the data.
+"""
 struct RawData <: Transformer end
+
 train!(::RawData, args...) = nothing
 StatsAPI.predict(::RawData, X) = X
 
-# z-score
-Base.@kwdef mutable struct ZScore <: Transformer
+"""
+    ZScore
+
+A transformer that scales and centers the data, using only the data that are
+avaiable to the model at training time.
+"""
+Base.@kwdef mutable struct  <: Transformer
     μ::AbstractArray = zeros(1)
     σ::AbstractArray = zeros(1)
 end
+
 function train!(zs::ZScore, X; kwdef...)
     zs.μ = vec(mean(X; dims = 2))
     zs.σ = vec(std(X; dims = 2))
